Finished implementing the raycast projection code
